HC-53: added ServerActionsService with tests

+++ source/Hetzner.Cloud/Interfaces/IServerService.cs
@@ -10,13 +10,13 @@ public interface IServerService
     /// <summary>
     /// Returns all existing Server objects
     /// </summary>
-    /// <param name="page">specifies the page to fetch. The number of the first page is 1.</param>
-    /// <param name="itemsPerPage">specifies the number of items returned per page. The default value is 25, the maximum value is 50 except otherwise specified in the documentation.</param>
+    /// <param name="page">Page to load.</param>
+    /// <param name="itemsPerPage">Items to load per page.</param>
     /// <param name="filter">filter allows only <seealso cref="NameFilter"/>, <seealso cref="LabelFilter"/> and <seealso cref="StatusFilter"/>.</param>
     /// <param name="sorting">sorts the result</param>
     /// <returns></returns>
-    Task<PagedResponse<Server>> GetAllAsync(int page = 1,
-        int itemsPerPage = 25,
+    Task<PagedResponse<Server>> GetAllAsync(long page = 1,
+        long itemsPerPage = 25,
         List<IFilter>? filter = null,
         Sorting<ServerSorting>? sorting = null,
         CancellationToken cancellationToken = default);
